requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
rembg==2.0.59
pillow==10.4.0
aiofiles==24.1.0
